Pontuação

+++ Jogo-Infinitas-Tentativas/Jogo-Infinitas-Tentativas.c
@@ -7,7 +7,7 @@ int main() {
     int numerosecreto = 42;
     int chute;
     int tentativa = 1;
-    int pontos = 1000;
+    double pontos = 1000;
 
 
     while(1){
@@ -39,12 +39,12 @@ int main() {
         }
 
         tentativa++;
-        int pontosperdidos = (chute - numerosecreto) / 2;
+        double pontosperdidos = (chute - numerosecreto) / 2.0;
         pontos = pontos - pontosperdidos;
     };
     printf("-----------------------------------\n");
     printf("Voce usou %d tentativas\n", tentativa);
     printf("FIM DE JOGO\n");
-    printf("Pontuacao: %d\n", pontos);
+    printf("Pontuacao: %.2f\n", pontos);
 
 };
